organize etrade components into well-named packages
