+++ src/main/kotlin/Samples.kt
@@ -1,12 +1,12 @@
 package com.seansoper.batil.samples
 
 import com.seansoper.batil.OptionsCalendar
-import com.seansoper.batil.brokers.etrade.Accounts
-import com.seansoper.batil.brokers.etrade.Alerts
-import com.seansoper.batil.brokers.etrade.Authorization
-import com.seansoper.batil.brokers.etrade.Market
-import com.seansoper.batil.brokers.etrade.Orders
-import com.seansoper.batil.brokers.etrade.TransactionSortOrder
+import com.seansoper.batil.brokers.etrade.auth.Authorization
+import com.seansoper.batil.brokers.etrade.services.Accounts
+import com.seansoper.batil.brokers.etrade.services.Alerts
+import com.seansoper.batil.brokers.etrade.services.Market
+import com.seansoper.batil.brokers.etrade.services.Orders
+import com.seansoper.batil.brokers.etrade.services.TransactionSortOrder
 import com.seansoper.batil.config.GlobalConfig
 import com.seansoper.batil.config.RuntimeConfig
 import java.util.GregorianCalendar
