+++ src/main/kotlin/brokers/etrade/auth/Session.kt
@@ -1,4 +1,4 @@
-package com.seansoper.batil.brokers.etrade
+package com.seansoper.batil.brokers.etrade.auth
 
 class Session(
     val consumerKey: String,
