+++ src/test/kotlin/AccountsTest.kt
@@ -1,15 +1,15 @@
 import TestHelper.MockHelper.createServer
 import TestHelper.MockHelper.mockSession
 import TestHelper.PathHelper.randomString
-import com.seansoper.batil.brokers.etrade.AccountMode
-import com.seansoper.batil.brokers.etrade.AccountType
-import com.seansoper.batil.brokers.etrade.Accounts
-import com.seansoper.batil.brokers.etrade.OptionLevel
-import com.seansoper.batil.brokers.etrade.OptionType
-import com.seansoper.batil.brokers.etrade.PositionIndicatorType
-import com.seansoper.batil.brokers.etrade.QuoteMode
-import com.seansoper.batil.brokers.etrade.SecurityType
-import com.seansoper.batil.brokers.etrade.TransactionSortOrder
+import com.seansoper.batil.brokers.etrade.api.AccountMode
+import com.seansoper.batil.brokers.etrade.api.AccountType
+import com.seansoper.batil.brokers.etrade.api.OptionLevel
+import com.seansoper.batil.brokers.etrade.api.OptionType
+import com.seansoper.batil.brokers.etrade.api.PositionIndicatorType
+import com.seansoper.batil.brokers.etrade.api.QuoteMode
+import com.seansoper.batil.brokers.etrade.api.SecurityType
+import com.seansoper.batil.brokers.etrade.services.Accounts
+import com.seansoper.batil.brokers.etrade.services.TransactionSortOrder
 import io.kotlintest.matchers.string.shouldContain
 import io.kotlintest.matchers.types.shouldBeNull
 import io.kotlintest.matchers.types.shouldNotBeNull
