+++ src/test/kotlin/AuthorizationTest.kt
@@ -2,7 +2,7 @@ import TestHelper.LoadConfig
 import TestHelper.MockHelper.createServer
 import TestHelper.PathHelper.randomString
 import com.seansoper.batil.brokers.etrade.AuthResponse
-import com.seansoper.batil.brokers.etrade.Authorization
+import com.seansoper.batil.brokers.etrade.auth.Authorization
 import io.kotlintest.matchers.string.shouldContain
 import io.kotlintest.matchers.types.shouldBeInstanceOf
 import io.kotlintest.matchers.types.shouldNotBeNull
