+++ src/test/kotlin/OrdersTest.kt
@@ -1,16 +1,16 @@
 import TestHelper.MockHelper.createServer
 import TestHelper.MockHelper.mockSession
 import TestHelper.PathHelper.randomString
-import com.seansoper.batil.brokers.etrade.MarginLevel
-import com.seansoper.batil.brokers.etrade.MessageType
-import com.seansoper.batil.brokers.etrade.OptionLevel
-import com.seansoper.batil.brokers.etrade.OptionType
-import com.seansoper.batil.brokers.etrade.OrderActionType
-import com.seansoper.batil.brokers.etrade.OrderStatus
-import com.seansoper.batil.brokers.etrade.OrderTerm
-import com.seansoper.batil.brokers.etrade.OrderType
-import com.seansoper.batil.brokers.etrade.Orders
-import com.seansoper.batil.brokers.etrade.SecurityType
+import com.seansoper.batil.brokers.etrade.api.MarginLevel
+import com.seansoper.batil.brokers.etrade.api.MessageType
+import com.seansoper.batil.brokers.etrade.api.OptionLevel
+import com.seansoper.batil.brokers.etrade.api.OptionType
+import com.seansoper.batil.brokers.etrade.api.OrderActionType
+import com.seansoper.batil.brokers.etrade.api.OrderTerm
+import com.seansoper.batil.brokers.etrade.api.OrderType
+import com.seansoper.batil.brokers.etrade.api.SecurityType
+import com.seansoper.batil.brokers.etrade.services.OrderStatus
+import com.seansoper.batil.brokers.etrade.services.Orders
 import io.kotlintest.matchers.string.shouldContain
 import io.kotlintest.matchers.types.shouldNotBeNull
 import io.kotlintest.shouldBe
